+++ log/example_test.go
@@ -14,7 +14,7 @@ import (
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-func Example_logger() {
+func ExampleNew() {
 	logger, err := New("/path/to/file.log", 0644)
 
 	if err != nil {
@@ -47,4 +47,340 @@ func Example_logger() {
 
 	// If buffered IO is used, you should flush data before exit
 	logger.Flush()
+}
+
+func Example_Reopen() {
+	err := Set("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	// For log rotation we provide method Reopen
+	Reopen()
+}
+
+func ExampleMinLevel() {
+	err := Set("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	// Set minimal log level to error
+	MinLevel(ERROR)
+
+	Info("This message is not displayed")
+	Error("This message is displayed")
+}
+
+func ExampleSet() {
+	err := Set("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	err = Set("/path/to/file2.log", 0640)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	Info("Message will go to file2.log")
+}
+
+func ExampleEnableBufIO() {
+	err := Set("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	// Enable buffered IO with 1-second flush interval
+	EnableBufIO(time.Second)
+
+	Info("Info message")
+
+	Flush()
+}
+
+func ExampleFlush() {
+	err := Set("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	// Enable buffered IO with 1-second flush interval
+	EnableBufIO(time.Second)
+
+	Info("Info message")
+
+	Flush()
+}
+
+func ExamplePrint() {
+	err := Set("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	Print(INFO, "Info message")
+	Print(ERROR, "Error message")
+}
+
+func ExampleDebug() {
+	err := Set("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	Debug("Debug message")
+}
+
+func ExampleInfo() {
+	err := Set("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	Info("Info message")
+}
+
+func ExampleWarn() {
+	err := Set("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	Warn("Warning message")
+}
+
+func ExampleError() {
+	err := Set("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	Error("Error message")
+}
+
+func ExampleCrit() {
+	err := Set("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	Crit("Critical error message")
+}
+
+func ExampleAux() {
+	err := Set("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	Aux("Auxiliary message")
+}
+
+func ExampleIs() {
+	err := Set("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	if Is(INFO) {
+		Info("Info message")
+	}
+}
+
+func ExampleLogger_Reopen() {
+	logger, err := New("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	// For log rotation we provide method Reopen
+	logger.Reopen()
+}
+
+func ExampleLogger_MinLevel() {
+	logger, err := New("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	// Set minimal log level to error
+	logger.MinLevel(ERROR)
+
+	logger.Info("This message is not displayed")
+	logger.Error("This message is displayed")
+}
+
+func ExampleLogger_Set() {
+	logger, err := New("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	err = logger.Set("/path/to/file2.log", 0640)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	logger.Info("Message will go to file2.log")
+}
+
+func ExampleLogger_EnableBufIO() {
+	logger, err := New("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	// Enable buffered IO with 1-second flush interval
+	logger.EnableBufIO(time.Second)
+
+	logger.Info("Info message")
+
+	logger.Flush()
+}
+
+func ExampleLogger_Flush() {
+	logger, err := New("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	// Enable buffered IO with 1-second flush interval
+	logger.EnableBufIO(time.Second)
+
+	logger.Info("Info message")
+
+	logger.Flush()
+}
+
+func ExampleLogger_Print() {
+	logger, err := New("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	logger.Print(INFO, "Info message")
+	logger.Print(ERROR, "Error message")
+}
+
+func ExampleLogger_Debug() {
+	logger, err := New("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	logger.Debug("Debug message")
+}
+
+func ExampleLogger_Info() {
+	logger, err := New("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	logger.Info("Info message")
+}
+
+func ExampleLogger_Warn() {
+	logger, err := New("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	logger.Warn("Warning message")
+}
+
+func ExampleLogger_Error() {
+	logger, err := New("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	logger.Error("Error message")
+}
+
+func ExampleLogger_Crit() {
+	logger, err := New("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	logger.Crit("Critical error message")
+}
+
+func ExampleLogger_Aux() {
+	logger, err := New("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	logger.Aux("Auxiliary message")
+}
+
+func ExampleLogger_Is() {
+	logger, err := New("/path/to/file.log", 0644)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	if logger.Is(INFO) {
+		logger.Info("Info message")
+	}
 }